@@ -2,17 +2,26 @@
 	import { cn } from '$lib/utils';
 	import { Motion } from 'svelte-motion';
 
-	export let word: string = 'Blur In';
-	export let variant: {
+	interface Props {
+		word?: string;
+		variant?: {
 		hidden: { filter: string; opacity: number };
 		visible: { filter: string; opacity: number };
-	} = {
+	};
+		duration?: number;
+		class?: string;
+	}
+
+	let {
+		word = 'Blur In',
+		variant = {
 		hidden: { filter: 'blur(10px)', opacity: 0 },
 		visible: { filter: 'blur(0px)', opacity: 1 }
-	};
-	export let duration: number = 0.75;
-	let className = '';
-	export { className as class };
+	},
+		duration = 0.75,
+		class: className = ''
+	}: Props = $props();
+	
 
 	let defaultVariants = {
 		hidden: { filter: 'blur(10px)', opacity: 0 },
