@@ -1,12 +1,24 @@
 <script lang="ts">
 	import { cn } from '$lib/utils';
-	export let pauseOnHover: boolean = false;
-	export let vertical: boolean = false;
-	export let repeat: number = 4;
-	export let reverse: boolean = false;
 
-	let className: any = '';
-	export { className as class };
+	interface Props {
+		pauseOnHover?: boolean;
+		vertical?: boolean;
+		repeat?: number;
+		reverse?: boolean;
+		class?: any;
+		children?: import('svelte').Snippet;
+	}
+
+	let {
+		pauseOnHover = false,
+		vertical = false,
+		repeat = 4,
+		reverse = false,
+		class: className = '',
+		children
+	}: Props = $props();
+	
 </script>
 
 <div
@@ -28,7 +40,7 @@
 				'[animation-direction:reverse]': reverse
 			})}
 		>
-			<slot>Default</slot>
+			{#if children}{@render children()}{:else}Default{/if}
 		</div>
 	{/each}
 </div>
