@@ -1,12 +1,25 @@
 <script lang="ts">
 	import { cn } from '$lib/utils';
-	let className: any = '';
-	export { className as class };
-	export let reverse: boolean = false;
-	export let duration: number = 20;
-	export let delay: number = 0;
-	export let radius: number = 50;
-	export let path: boolean = true;
+	
+	interface Props {
+		class?: any;
+		reverse?: boolean;
+		duration?: number;
+		delay?: number;
+		radius?: number;
+		path?: boolean;
+		children?: import('svelte').Snippet;
+	}
+
+	let {
+		class: className = '',
+		reverse = false,
+		duration = 20,
+		delay = 0,
+		radius = 50,
+		path = true,
+		children
+	}: Props = $props();
 </script>
 
 {#if path}
@@ -34,6 +47,6 @@
 			className
 		)}
 	>
-		<slot></slot>
+		{@render children?.()}
 	</div>
 {/if}
