@@ -1,7 +1,6 @@
 <script lang="ts">
 	import * as Form from '$lib/components/ui/form';
 	import { fade } from 'svelte/transition';
-	export let data;
 	import { Input } from '$lib/components/ui/input';
 	import { registerUserSchema, type RegisterUserSchema } from '$lib/schema';
 	import { type SuperValidated, type Infer, superForm } from 'sveltekit-superforms';
@@ -11,8 +10,13 @@
 	import { toast } from 'svelte-sonner';
 	import { onDestroy, onMount, tick } from 'svelte';
 	import { siteInfo } from '$lib/data.js';
+	interface Props {
+		data: any;
+	}
 
-	let isSubmitting = false;
+	let { data }: Props = $props();
+
+	let isSubmitting = $state(false);
 
 	const form = superForm(data.form, {
 		validators: zodClient(registerUserSchema)
